% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarity.R
\name{calc_rarity}
\alias{calc_rarity}
\title{Calculate the Index of Relative Rarity for Species Assemblages}
\usage{
calc_rarity(Fin, RW)
}
\arguments{
\item{Fin}{A data frame of final species presences, containing columns for \code{cell}, \code{species}, and \code{Landuse}.}

\item{RW}{A data frame of rarity weights, as calculated by the \code{calc_rarity_weight} function.}
}
\value{
A data frame containing the Index of Relative Rarity (Irr) for each cell, along with the corresponding land-use type.
}
\description{
This function calculates the Index of Relative Rarity (Irr) for species assemblages based on their occurrences in specific land-use types. It uses the rarity weights calculated from the \code{calc_rarity_weight} function.
}
