% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Taxa.R
\name{Clean_Taxa}
\alias{Clean_Taxa}
\title{Clean taxa using Taxize and rgbif}
\usage{
Clean_Taxa(Taxons, WriteFile = FALSE, Species_Only = TRUE, verbose = TRUE)
}
\arguments{
\item{Taxons}{Vector of taxa to be cleaned.}

\item{WriteFile}{logical; if FALSE (default) only returns a data frame, if TRUE will generate a folder (Results) with CSVs.}

\item{Species_Only}{logical; if TRUE (default) only species will be returned, if FALSE returns highest possible taxonomic resolution.}

\item{verbose}{logical; if TRUE, prints progress and summary messages (default: TRUE).}
}
\value{
A data frame with the cleaned taxa and their scores.
}
\description{
This function cleans a vector of taxa using Taxize and rgbif sequentially.
}
\examples{
Cleaned <- Clean_Taxa(c("Canis lupus", "C. lupus"), verbose = TRUE)
}
