% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Taxa.R
\name{Clean_Taxa_rgbif}
\alias{Clean_Taxa_rgbif}
\title{Clean Taxa from GBIF}
\usage{
Clean_Taxa_rgbif(
  Cleaned_Taxize,
  WriteFile = FALSE,
  Species_Only = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{Cleaned_Taxize}{a data frame containing the cleaned taxonomic list from \code{Clean_Taxa_Taxize}.}

\item{WriteFile}{logical; if FALSE (default) only returns a data frame, if TRUE will generate a folder (Results) with a CSV.}

\item{Species_Only}{logical; if TRUE (default) only species will be returned, if FALSE returns highest possible taxonomic resolution.}

\item{verbose}{logical; if TRUE, prints progress and summary messages (default: FALSE).}
}
\value{
A data frame containing the GBIF-cleaned taxonomic list.
}
\description{
Clean the taxonomic list using GBIF.
}
\examples{
\dontrun{
Cleaned_Taxize <- Clean_Taxa_Taxize(c("Abies concolor", "Canis lupus"), verbose = TRUE)
Clean_Taxa_rgbif(Cleaned_Taxize, verbose = TRUE)
}
}
