% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Taxa.R
\name{Clean_Taxa_Taxize}
\alias{Clean_Taxa_Taxize}
\title{Clean taxa using Taxize}
\usage{
Clean_Taxa_Taxize(Taxons, WriteFile = FALSE, verbose = TRUE)
}
\arguments{
\item{Taxons}{Vector of taxa to be cleaned.}

\item{WriteFile}{logical; if FALSE (default) only returns a
data frame, if TRUE will generate a folder (Results) in the
working directory with a CSV of the results.}

\item{verbose}{logical; if TRUE, prints progress and summary
messages (default: TRUE).}
}
\value{
A data frame with the cleaned taxa and their scores.
}
\description{
This function cleans a vector of taxa using the Global Names Architecture
via the \pkg{taxize} package.
}
\examples{
Clean_Taxa_Taxize(
  Taxons =  c("Abies concolor", "Abies lowiana", "Canis lupus", "Cannis lupus"),
  verbose = TRUE
)

}
