% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM.R
\name{Convex_20}
\alias{Convex_20}
\title{Convex_20}
\usage{
Convex_20(
  DF,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  proj = "+proj=longlat +datum=WGS84 +no_defs",
  expansion = 1.2
)
}
\arguments{
\item{DF}{The dataframe containing the coordinates.}

\item{lon}{The name of the longitude column in the dataframe.}

\item{lat}{The name of the latitude column in the dataframe.}

\item{proj}{The projection of the coordinates.}

\item{expansion}{Numeric expansion factor applied to the convex hull
around its centroid. Default 1.2 (a 20\% expansion).}
}
\value{
A polygon representing the expanded convex hull.
}
\description{
Creates an expanded convex hull from a set of coordinates, used to define
a background-sampling area around a species' known occurrences.
}
\examples{

DF <- data.frame(decimalLongitude =
      c(23.978543, 23.785003, 11.485,  -2.054027, 12.9069),
                 decimalLatitude =
       c(38.088876, 60.238213, 48.165, 53.33939, 56.80782))

Convex_20(DF, lon = "decimalLongitude", lat = "decimalLatitude",
proj = "+proj=longlat +datum=WGS84 +no_defs")

}
