% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_final_presences.R
\name{export_final_presences}
\alias{export_final_presences}
\title{Export Final Species Presences to CSV Files}
\usage{
export_final_presences(DF, folder)
}
\arguments{
\item{DF}{A data frame containing the final species presences with columns including \code{species} and \code{Landuse}.}

\item{folder}{A string specifying the directory where the CSV files should be saved. The function will create subfolders for each land-use type within this directory.}
}
\description{
This function exports the final filtered species presences to CSV files. The data is grouped by land-use type, and each species is saved in a separate CSV file within a folder corresponding to its land-use type.
}
