% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pd.R
\name{export_pd}
\alias{export_pd}
\title{Export Phylogenetic Diversity to Cloud-Optimized GeoTIFF}
\usage{
export_pd(Results, path)
}
\arguments{
\item{Results}{A data.table containing Phylogenetic Diversity results. This data.table must contain at least the columns
\code{cell} and \code{PD}, where \code{cell} refers to the cell index in the raster and \code{PD} refers to the Phylogenetic Diversity value.}

\item{path}{A character string specifying the file path to a raster template. This raster is used as a template for the output raster.}
}
\description{
This function takes a data.table of Phylogenetic Diversity results and a path to a raster template.
It generates a \code{SpatRaster} object using the \code{terra} package, populates it with the Phylogenetic Diversity values,
and writes the result out as a Cloud-Optimized GeoTIFF.
}
