% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_rarity.R
\name{export_rarity}
\alias{export_rarity}
\title{Export Rarity to Cloud-Optimized GeoTIFF}
\usage{
export_rarity(Results, path)
}
\arguments{
\item{Results}{A data.table containing rarity results. This data.table must contain at least the columns
\code{cell} and \code{Irr}, where \code{cell} refers to the cell index in the raster and \code{Irr} refers to the rarity value.}

\item{path}{A character string specifying the file path to a raster template. This raster is used as a template for the output raster.}
}
\value{
A character string indicating the file path of the output Cloud-Optimized GeoTIFF.
}
\description{
This function takes a data.table of rarity results and a path to a raster template.
It generates a \code{SpatRaster} object using the \code{terra} package, populates it with the rarity values,
and writes the result out as a Cloud-Optimized GeoTIFF.
}
