% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_valid_presences.R
\name{filter_valid_presences}
\alias{filter_valid_presences}
\title{Filter Valid Species Presences}
\usage{
filter_valid_presences(x)
}
\arguments{
\item{x}{A data frame or list of data frames returned by
\code{\link[=get_presences]{get_presences()}}.}
}
\value{
A list of cleaned, non-empty presence data frames.
}
\description{
Removes malformed, empty, duplicated, and spatially invalid occurrence
datasets.
}
