% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_landuse_table.R
\name{generate_long_landuse_table}
\alias{generate_long_landuse_table}
\title{Generate and Transform a Land-Use Suitability Table}
\usage{
generate_long_landuse_table(path)
}
\arguments{
\item{path}{A string representing the file path to the raster file containing land-use suitability data.}
}
\value{
A data frame in long format with the raster cells that have a suitability value of 1 in at least one of the land-use categories. The data frame includes the \code{cell} identifier and the corresponding habitat types (\code{Habitat}) where suitability is 1.
}
\description{
This function takes the path to a raster file containing land-use suitability values, converts it to a data frame, and filters it to include only the cells that have a value of 1 in at least one of the land-use categories. It then transforms the filtered data into a long-format table where each row corresponds to a specific cell and habitat type.
}
\examples{
# Get path for habitat suitability raster
HabSut <- system.file("ex/HabSut.tif", package = "SpeciesPoolR")

# Use the function to generate and transform the land-use suitability table
generate_long_landuse_table(path = HabSut)

}
