% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM.R
\name{Generate_Lookup}
\alias{Generate_Lookup}
\title{Generate a Lookup Table for Species Land-Use Preferences}
\usage{
Generate_Lookup(Model, Thresholds)
}
\arguments{
\item{Model}{A data frame containing model predictions, with columns for \code{species}, \code{Landuse}, and \code{Pred}.}

\item{Thresholds}{A data frame containing the thresholds for presence predictions, with columns for \code{species}, \code{Thres_99}, \code{Thres_95}, and \code{Thres_90}.}
}
\value{
A data frame that indicates the land-use types where species are predicted to be present. The data frame contains columns for \code{species}, \code{Landuse}, and \code{Pres}.
}
\description{
This function generates a lookup table that indicates the land-use types where species are predicted to be present, based on model predictions and specified thresholds.
}
\examples{
lookup_table <- Generate_Lookup(Model = model_output, Thresholds = thresholds)

}
