% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read in Data from a CSV or XLSX File}
\usage{
get_data(file, filter = NULL)
}
\arguments{
\item{file}{A string specifying the path to the CSV or XLSX file to read. The file must contain a "Species" column.}

\item{filter}{An optional expression used to filter the resulting \code{data.frame}. This should be an expression
written as if you were using \code{dplyr::filter()}. The default is NULL, meaning no filtering is applied.}
}
\value{
A \code{data.frame} containing the data read from the file, optionally filtered.
}
\description{
This function reads in data from a CSV or XLSX file. The file must contain a "Species" column.
You can optionally apply a filter expression to the dataset before returning it.
The function returns a \code{data.frame} with the filtered results for further processing in the pipeline.
}
\examples{
\dontrun{
# Read in data without filtering
f <- system.file("ex/Species_List.csv", package="SpeciesPoolR")
data <- get_data(f)

# Read in data and filter for Plantae kingdom and certain taxon ranges
filtered_data <- get_data(
  file = f,
  filter = quote(
   Kingdom == "Plantae" &
   Class == "Magnoliopsida"
  )
)
}
}
