% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presences.R
\name{get_presences}
\alias{get_presences}
\title{Get Occurrence Data for a Given Species List}
\usage{
get_presences(species, country = NULL, shapefile = NULL)
}
\arguments{
\item{species}{A vector of species to use.}

\item{country}{A two-letter country code (e.g., "DK" for Denmark) to define the area of interest.}

\item{shapefile}{A shapefile (with lat/long coordinates) defining the area of interest.
The function will create a minimum bounding rectangle around the shapefile to query the
species occurrences. Default is \code{NULL}.}
}
\value{
A data.frame containing the occurrence data for the specified species, including the columns: \code{scientificName}, \code{decimalLatitude}, \code{decimalLongitude}, \code{family}, \code{genus}, and \code{species}.
}
\description{
This function retrieves occurrence data for a given list of species using the \code{GetOccs} function. The function is designed to fetch data for species in Denmark (\code{country = "DK"}) from 1999 to 2023 and return a cleaned data frame with selected columns.
}
\examples{
\dontrun{
# Assume `species_list` is a data.frame with a column named `species`
presences <- get_presences(species_list)
}
}
