% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLandusePresences.R
\name{GetLandusePresences}
\alias{GetLandusePresences}
\title{Get Landuse Presences}
\usage{
GetLandusePresences(folder, Landuse)
}
\arguments{
\item{folder}{A character string specifying the path to the folder containing the land use data files.}

\item{Landuse}{A character string specifying the land use type for which the presence data is to be read. The function assumes that the files are located in subdirectories named after the land use type within the specified folder.}
}
\value{
A data frame containing combined species presence data with columns \code{cell}, \code{species}, and \code{Landuse}.
}
\description{
This function reads and combines species presence data from multiple files within a specified folder for a given land use type.
}
\examples{
\dontrun{
  # Example usage:
  folder <- system.file("ex/", package="SpeciesPoolR")
  Landuse <- "ForestDryPoor"
  presences <- GetLandusePresences(folder, Landuse)
  head(presences)
}
}
