% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presences.R
\name{GetOccs}
\alias{GetOccs}
\title{Get occurrence data}
\usage{
GetOccs(
  Species,
  WriteFile = T,
  continent = NULL,
  country = NULL,
  shapefile = NULL,
  limit = 10000,
  Log = T,
  ...
)
}
\arguments{
\item{Species}{A vector containing the species to query.}

\item{WriteFile}{Logical. If \code{TRUE}, the occurrence data will be written to the \code{Occs} folder. If \code{FALSE}, the occurrence data will be returned in a list.}

\item{continent}{what contintent are the occurrences downloaded from}

\item{country}{A two-letter country code (e.g., "DK" for Denmark) to define the area of interest.}

\item{shapefile}{A shapefile (with lat/long coordinates) defining the area of interest.
The function will create a minimum bounding rectangle around the shapefile to query the
species occurrences. Default is \code{NULL}.}

\item{limit}{maximum number of occurrences downloaded}

\item{Log}{Logical. If \code{TRUE}, a log file will be created with information on the progress of the function.}

\item{...}{Additional arguments to be passed to the \code{occ_data} function.}
}
\value{
If \code{WriteFile = TRUE}, this function does not return anything. If \code{WriteFile = FALSE}, a list containing the occurrence data for each species is returned.
}
\description{
This function uses the \code{rgbif} package to get occurrence data from the Global Biodiversity Information Facility (GBIF) API.
}
\examples{
# Get occurrence data for species in FinalSpeciesList
\donttest{
Presences <- GetOccs(Species = c("Abies concolor", "Canis lupus"), WriteFile = FALSE)
}
}
