% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Richness_PD}
\alias{Richness_PD}
\title{Species Richness and Phylogenetic Diversity Data}
\format{
A data.table and data.frame with 26,433 rows and 4 variables:
\describe{
\item{PD}{Numeric. The Phylogenetic Diversity calculated with the \code{picante} R package.}
\item{SR}{Integer. The Species Richness calculated with the \code{picante} R package.}
\item{cell}{Integer. The cell index in the raster.}
\item{Landuse}{Character. The type of land use associated with the cell (e.g., "ForestDryPoor").}
}
}
\usage{
Richness_PD
}
\description{
A dataset containing species richness and phylogenetic diversity results along with associated land use information.
}
\examples{
\dontrun{
  data(Richness_PD)
  head(Richness_PD)
}
}
\keyword{datasets}
