% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Richness_Rar}
\alias{Richness_Rar}
\title{Species Richness and Rarity Data}
\format{
A data.table and data.frame with 26,433 rows and 4 variables:
\describe{
\item{cell}{Numeric. The cell index in the raster.}
\item{Irr}{Numeric. The Index of Relative Rarity calculated with the \code{rarity} R package.}
\item{Richness}{Integer. The species richness value for the cell.}
\item{Landuse}{Character. The type of land use associated with the cell (e.g., "ForestDryPoor").}
}
}
\usage{
Richness_Rar
}
\description{
A dataset containing species richness results along with associated Index of Relative Rarity (Irr) and land use information.
}
\examples{
\dontrun{
  data(Richness_Rar)
  head(Richness_Rar)
}
}
\keyword{datasets}
