% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM.R
\name{SampleLanduse}
\alias{SampleLanduse}
\title{Sample Land-Use Data for Species Presences or Background Locations}
\usage{
SampleLanduse(DF, file, type = "pres")
}
\arguments{
\item{DF}{A data frame containing species presence or background data, with columns for \code{species}, \code{decimalLongitude}, and \code{decimalLatitude}.}

\item{file}{A string representing the path to a raster file that contains land-use data.}

\item{type}{A string specifying whether to sample land-use data from species presences (\code{"pres"}) or from background locations (\code{"bg"}). Defaults to \code{"pres"}.}
}
\value{
A data frame with the sampled land-use data. The data frame contains columns for \code{species}, \code{Landuse}, and \code{Pres}, where \code{Pres} indicates whether the row corresponds to a presence (1) or background (0) point.
}
\description{
This function samples land-use data either from species presence locations or from background locations within a specified geographic area.
}
