% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarity.R
\name{calc_rarity_weight}
\alias{calc_rarity_weight}
\title{Calculate Rarity Weights for Species}
\usage{
calc_rarity_weight(df)
}
\arguments{
\item{df}{A data frame containing at least two columns: \code{species} and \code{N}, where \code{species} represents the species names and \code{N} represents the number of occurrences for each species.}
}
\value{
A data frame with rarity weights for each species, calculated using the \code{rWeights} function.
}
\description{
This function calculates rarity weights for a given set of species occurrences using the \code{rWeights} function from the \code{Rarity} package. The weights are based on the species' occurrences and are used to assess their rarity within the dataset.
}
