% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM.R
\name{create_thresholds}
\alias{create_thresholds}
\title{Create Prediction Thresholds for Species Distribution Models}
\usage{
create_thresholds(Model, reference, file)
}
\arguments{
\item{Model}{A data frame containing model predictions, with columns for \code{species}, \code{Landuse}, and \code{Pred}.}

\item{reference}{A data frame containing reference species presence data for threshold calibration.}

\item{file}{A string representing the path to a raster file that contains land-use data.}
}
\value{
A data frame with the calculated thresholds. The data frame contains columns for \code{species}, \code{Thres_99}, \code{Thres_95}, and \code{Thres_90}.
}
\description{
This function generates thresholds for species distribution predictions based on modeled land-use preferences. Thresholds are calculated for the 99th, 95th, and 90th percentiles of the predicted values.
}
\examples{
thresholds <- create_thresholds(Model = model_output, reference = reference_data, file = file)

}
