% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_tree}
\alias{generate_tree}
\title{Generate Phylogenetic Tree}
\usage{
generate_tree(DF)
}
\arguments{
\item{DF}{A character string specifying the file path to a CSV file. The CSV file must contain columns named \code{species}, \code{genus}, and \code{family}.}
}
\value{
A phylogenetic tree object generated by the \code{V.PhyloMaker::phylo.maker} function.
}
\description{
This function takes a CSV file path, reads the data, and generates a phylogenetic tree using the \code{V.PhyloMaker} package.
}
\examples{
\dontrun{
  # Example usage:
  p <- system.file("ex/Clean_Species.csv", package="SpeciesPoolR")
  tree <- generate_tree(p)

}
}
