% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM.R
\name{ModelAndPredictFunc}
\alias{ModelAndPredictFunc}
\title{Model and Predict Habitat Suitability}
\usage{
ModelAndPredictFunc(DF, file)
}
\arguments{
\item{DF}{A data frame containing species presence data with columns for species name, longitude (\code{decimalLongitude}), and latitude (\code{decimalLatitude}).}

\item{file}{A file path to the raster layer containing land-use data.}
}
\value{
A data frame with predicted habitat suitability scores for each land-use type for each species.
}
\description{
This function performs the complete workflow for modeling habitat suitability for multiple species. It includes sampling land-use data for presence and background points separately, and then fitting a MaxEnt model to predict habitat suitability based on the available land-use types.
}
\details{
The function encompasses several steps:
\enumerate{
\item Grouping the data by species using \code{dplyr::group_split}.
\item Sampling land-use data for species presence points using the \code{SampleLanduse} function.
\item Sampling land-use data for background points using the \code{SampleLanduse} function.
\item Combining the presence and background data, and fitting a MaxEnt model to predict habitat suitability using the \code{ModelSpecies} function, which also handles the duplication of rows where necessary.
}
}
