% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM.R
\name{ModelSpecies}
\alias{ModelSpecies}
\title{Fit a Species Distribution Model Based on Land-Use Data}
\usage{
ModelSpecies(DF)
}
\arguments{
\item{DF}{A data frame containing species presence and background data, with columns for \code{species}, \code{Landuse}, and \code{Pres}. The \code{Landuse} column should be a factor representing different land-use types, and \code{Pres} should indicate whether the row corresponds to a presence (1) or background (0) point.}
}
\value{
A data frame with predicted species distribution for each land-use type. The data frame contains columns for \code{Landuse}, \code{Pred} (predicted value), and \code{species}.
}
\description{
This function fits a MaxEnt model to predict species distribution based on land-use data. It handles cases where land-use types are listed as "Both" by splitting these into separate "Poor" and "Rich" categories.
}
\examples{
model_output <- ModelSpecies(DF = combined_data)

}
