% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Run the SpeciesPoolR Workflow}
\usage{
run_workflow(
  workers = 2,
  error = "null",
  file_path,
  rastertemp,
  rasterLU,
  LanduseSuitability,
  dist = 500,
  filter = NULL,
  country = NULL,
  shapefile = NULL,
  plot = TRUE
)
}
\arguments{
\item{workers}{Number of parallel workers to use in the \code{crew} controller. Default is 2.}

\item{error}{Handling for errors in outdated targets. Default is "null".}

\item{file_path}{Path to the Excel or csv file containing the data.}

\item{rastertemp}{A file path to the raster file that will be used as a template for rasterizing the buffers.}

\item{rasterLU}{A file path to the raster file that has the landuses that exist in the area that will be modeled.}

\item{LanduseSuitability}{A string representing the file path to the raster file containing land-use binary suitability data.}

\item{dist}{A numeric value specifying the buffer distance in meters. Default is 500 meters.}

\item{filter}{An optional expression used to filter the resulting \code{data.frame}. This should be an expression
written as if you were using \code{dplyr::filter()}. The default is NULL, meaning no filtering is applied.}

\item{country}{A two-letter country code to define the area of interest for counting species presences. Default is NULL.}

\item{shapefile}{Path to a shapefile defining the area of interest for counting species presences. Default is NULL.}

\item{plot}{if TRUE (default) it will run the \code{targets::tar_visnetwork()} to plot the workflow}
}
\value{
Executes the \code{targets} pipeline.
}
\description{
This function sets up and runs a \code{targets} workflow using the functions provided in the
\code{SpeciesPoolR} package. The workflow includes steps for cleaning species data, counting
species presences, filtering data, and generating buffers for spatial analysis.
}
