% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Run the SpeciesPoolR Workflow}
\usage{
run_workflow(workers = 2, error = "null", file_path)
}
\arguments{
\item{workers}{Number of parallel workers to use in the \code{crew} controller. Default is 2.}

\item{error}{Handling for errors in outdated targets. Default is "null".}

\item{file_path}{Path to the Excel or csv file containing the data.}
}
\value{
Executes the \code{targets} pipeline.
}
\description{
This function sets up and runs a \code{targets} workflow using the functions provided in the
\code{SpeciesPoolR} package. The workflow includes steps for cleaning species data, counting
species presences, filtering data, and generating buffers for spatial analysis.
}
