% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cog.R
\name{write_cog}
\alias{write_cog}
\title{Write Cloud Optimized Geotiff (COG)}
\usage{
write_cog(SpatRaster, Name)
}
\arguments{
\item{SpatRaster}{A SpatRaster object (class SpatRaster) representing the raster data to be saved as a COG.}

\item{Name}{The desired name for the COG file, including the ".tif" extension.}
}
\value{
A Cloud Optimized Geotiff saved at the specified location.
}
\description{
This function takes a SpatRaster object and saves it as a Cloud Optimized Geotiff (COG).
COGs are geospatial files optimized for efficient cloud storage and retrieval.
}
\examples{
# Load required libraries if not already loaded
# library(terra)

# Create a sample SpatRaster
r <- terra::rast(nrows = 5, ncols = 5, vals = 1:25)

# Save the SpatRaster as a COG
write_cog(SpatRaster = r, Name = "test.tif")

# Clean up later
file.remove("test.tif")
file.remove("test.tfw")

}
